file directory start
This is the beginnings of a GUI for the project and functions as a file directory and text document reader. With this we have the have the means to find the pathway to a file and share that information with our AI program. Will post a PNG and Jpeg version later today.